Reverting some refactoring that'll make the merge hard
