--- soos-client/src/routes/App.tsx
@@ -1,9 +1,0 @@
-import { Entry } from './Entry';
-
-export function App() {
-  return (
-    <div>
-      <Entry />
-    </div>
-  )
-}
